+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added `cleanup_indexes` command to clean up outdated indexes and inaccessible repositories.
 
+### Fixed
+
+- Fixed connection closed or lost on `ConnectionPool` by using `check_connection` to verify if the connection is still working.
+
 ## [0.1.4] - 2025-05-22
 
 ### Added