+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.1.5] - 2025-05-26
+
 ### Added
 
 - Added `cleanup_indexes` command to clean up outdated indexes and inaccessible repositories.
@@ -638,7 +640,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of the `daiv` project.
 
-[Unreleased]: https://github.com/srtab/daiv/compare/v0.1.4...HEAD
+[Unreleased]: https://github.com/srtab/daiv/compare/v0.1.5...HEAD
+[0.1.5]: https://github.com/srtab/daiv/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/srtab/daiv/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/srtab/daiv/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/srtab/daiv/compare/v0.1.1...v0.1.2